update serialization/dubbo-samples-prefer-serialization module
