+++ 3-extensions/serialization/dubbo-samples-prefer-serialization/dubbo-samples-prefer-serialization-provider/src/main/java/org/apache/dubbo/samples/prefer/serialization/EmbeddedZooKeeper.java
